+++ .gitignore
@@ -3,5 +3,5 @@ pkg
 bin
 bundle
 
-# Ignore external dependencies (go get ./...)
-src/gopkg.in/
+# go build . output
+azure-docker-extension
