+++ Makefile
@@ -24,7 +24,7 @@ bundle: clean binary
 	@mkdir -p $(BUNDLEDIR)
 	zip ./$(BUNDLEDIR)/$(BUNDLE) ./$(BINDIR)/$(BIN)
 	zip -j ./$(BUNDLEDIR)/$(BUNDLE) ./metadata/HandlerManifest.json
-	zip ./$(BUNDLEDIR)/$(BUNDLE) ./util/run-in-background.sh
+	zip ./$(BUNDLEDIR)/$(BUNDLE) ./scripts/run-in-background.sh
 	@echo "OK: Use $(BUNDLEDIR)/$(BUNDLE) to publish the extension."
 binary:
 	GOPATH=`pwd` go get -v ./...
