+++ metadata/HandlerManifest.json
@@ -3,7 +3,7 @@
   "version": 1.0,
   "handlerManifest": {
     "installCommand":   "bin/docker-extension install",
-    "enableCommand":    "util/run-in-background.sh enable",
+    "enableCommand":    "scripts/run-in-background.sh enable",
     "uninstallCommand": "bin/docker-extension uninstall",
     "updateCommand":    "bin/docker-extension update",
     "disableCommand":   "bin/docker-extension disable",
