+++ op_disable.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"docker-extension/driver"
-	"vmextension"
+	"github.com/Azure/azure-docker-extension/pkg/driver"
+	"github.com/Azure/azure-docker-extension/pkg/vmextension"
 )
 
 func disable(he vmextension.HandlerEnvironment, d driver.DistroDriver) error {
