+++ op_enable_test.go
