+++ pkg/vmextension/handlersettings_test.go
