+++ pkg/vmextension/seqnum.go
