+++ pkg/vmextension/seqnum_test.go
